Refactored highlighting logic in graph component
