+++ client/src/pages/DepthFirstSearch/index.tsx
@@ -10,11 +10,25 @@ import AlgorithmPopover from "../../components/AlgorithmPopover";
 import { resetSteps, updateAlgorName } from "../../redux/stateSlice";
 import { depthFirstSearchDesc } from "../../assets/algorithm-information.js";
 import Stack from "../../components/Stack";
+import { RootState } from "../../redux/configureStore";
+import { GraphAlgorithmResultType } from "../../AlgoResultTypes";
+import { Edge } from "../../CommonTypes";
 
 const ALGORITHM_URL = "searches/depthfirstsearch/";
 
 const DepthFirstSearch = () => {
+
     const dispatch = useDispatch();
+    const n = 2000;
+
+    const algorSteps = useSelector(
+        (state: RootState) => state.global.algorSteps
+    ) as GraphAlgorithmResultType;
+    const currentStep = useSelector(
+        (state: RootState) => state.global.currentStep
+    );
+
+
     // reset data upon exiting the page
     useEffect(() => {
         // update the name on first load
@@ -25,7 +39,36 @@ const DepthFirstSearch = () => {
         };
     }, []);
 
-    const n = 2000;
+    const nodeHighlightStyle = (id: string) => {
+        let style = " ";
+        if (currentStep < 1 || algorSteps.steps.length === 0) return style;
+
+        if (algorSteps.steps[currentStep - 1].currentNode.includes(id)) {
+            style += "node-active ";
+        } else if (
+            algorSteps.steps[currentStep - 1].visitedNodes.includes(id)
+        ) {
+            style += "node-highlighted ";
+        }
+        return style;
+    };
+
+    const edgeHighlightStyle = (edge: Edge) => {
+        let style = " ";
+        if (currentStep < 1 || algorSteps.steps.length === 0) return style;
+
+        let currentEdgeList = algorSteps.steps[currentStep - 1].visitedEdges;
+        for (const edg of currentEdgeList) {
+            if (`${edge.n1} ${edge.n2}` === `${edg.n1} ${edg.n2}`) {
+                style += "edge-highlighted ";
+                break;
+            } else if (`${edge.n1} ${edge.n2}` === `${edg.n2} ${edg.n1}`) {
+                style += "edge-highlighted-reverse ";
+                break;
+            }
+        }
+        return style;
+    };
 
     return (
         <div className="content">
@@ -38,7 +81,12 @@ const DepthFirstSearch = () => {
                 initPosition={{ x: 0, y: 0 }}
                 staticChildren={<Stack />}
             >
-                <Graph containerWidth={n} containerHeight={n} />
+                <Graph
+                    containerWidth={n}
+                    containerHeight={n}
+                    edgeHighlightStyle={edgeHighlightStyle}
+                    nodeHighlightStyle={nodeHighlightStyle}
+                />
             </VisualizerContainer>
 
             <Controls algorithmUrl={ALGORITHM_URL} require={["graphInput"]} />
